menambahkan drawer button dan memindahkan file utils, views ke scripts
