+++ starter-project-harma/src/scripts/index.js
@@ -1,7 +1,7 @@
 import 'regenerator-runtime';
 import '../styles/main.css';
 import '../styles/responsive.css';
-import App from '../views/app';
+import App from './views/app';
 
 console.log('Hello Coders! :)');
 
@@ -9,4 +9,4 @@ const app = new App({
     button: document.querySelector('#hamburgerButton'),
     drawer: document.querySelector('#navigationDrawer'),
     content: document.querySelector('#mainContent'),
-});
+  });
