+++ starter-project-harma/src/scripts/views/app.js
@@ -1,0 +1,21 @@
+import DrawerInitiator from '../utils/drawer-initiator';
+ 
+class App {
+  constructor({ button, drawer, content }) {
+    this._button = button;
+    this._drawer = drawer;
+    this._content = content;
+ 
+    this._initialAppShell();
+  }
+ 
+  _initialAppShell() {
+    DrawerInitiator.init({
+      button: this._button,
+      drawer: this._drawer,
+      content: this._content,
+    });
+  }
+}
+ 
+export default App;
