+++ starter-project-harma/src/styles/main.css
@@ -22,34 +22,91 @@ body {
   display: block;
 }
 
-.app-bar__navigation {
-  display: none;
-}
-
-.app-bar__navigation ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.app-bar__navigation ul li {
-  padding: 6px;
-}
-
-.app-bar__navigation ul li a {
-  text-decoration: none;
-  color: #2A6478;
-  font-weight: bold;
-}
-
 .app-bar__navigation ul li a.login {
   background-color: #2A6478;
   color: #FFFFFF;
   padding: 8px 16px;
   border-radius: 4px;
+}
+
+.app-bar {
+  padding: 8px 16px;
+  background-color: #FEF4F4;
+  display: flex; 
+  justify-content: space-between; 
+  align-items: center;
+  position: sticky;
+  top: 0;
+  z-index: 99;
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.2);
+}
+ 
+.app-bar .app-bar__menu {
+  display: flex;
+  align-items: center;
+}
+ 
+.app-bar .app-bar__menu button {
+  background-color: transparent;
+  border: none;
+  color: black;
+  font-size: 25px;
+  padding: 12px;
+  cursor: pointer;
+}
+ 
+.app-bar .app-bar__brand {
+  display: flex;
+  align-items: center;
+}
+  
+.app-bar .app-bar__navigation {
+  position: absolute;
+  top: 50px;
+  left: -180px;
+  width: 150px;
+  transition: all 0.3s;
+  padding: 8px;
+}
+ 
+.app-bar .app-bar__navigation.open {
+  position: absolute;
+  top: 50px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 100%;
+  transition: all 0.3s;
+  padding: 8px;
+  overflow: hidden;
+}
+
+.app-bar .app-bar__navigation ul {
+  background-color: #FEF4F4;
+  border-radius: 10px;
+  list-style: none;
+  text-align: center;
+}
+
+.app-bar .app-bar__navigation ul li {
+  padding: 6px;
+  text-decoration: none;
+}
+
+.app-bar .app-bar__navigation ul li a {
+  display: inline-block;
+  text-decoration: none;
+  color: black;
+  padding: 8px;
+  margin-bottom: 5px;
+  width: 100%;
+  text-align: center;
+  transition: background-color 0.3s;
+}
+
+.app-bar .app-bar__navigation ul li a:hover {
+  background-color: #2A6478;
+  color: #1a120b;
+  border-radius: 8px;
 }
 
 .hero {
@@ -86,36 +143,4 @@ body {
   background-color: #1A4756;
   height: 100px;
   border-radius: 8px;
-}
-
-@media screen and (min-width: 650px) {
-  .app-bar__menu {
-      display: none;
-  }
-
-  .app-bar__navigation {
-      display: block;
-  }
-
-  .app-bar__navigation ul {
-      flex-direction: row;
-      justify-content: flex-end;
-      gap: 20px;
-  }
-
-  .app-bar__navigation ul li {
-      padding: 0;
-  }
-
-  .app-bar__navigation ul li a {
-      background-color: transparent;
-      color: #2A6478;
-  }
-
-  .app-bar__navigation ul li a.login {
-      background-color: #2A6478;
-      color: #FFFFFF;
-      padding: 8px 16px;
-      border-radius: 4px;
-  }
 }
