--- starter-project-harma/src/utils/drawer-initiator.js
@@ -1,23 +1,0 @@
-const DrawerInitiator = {
-    init({ button, drawer, content }) {
-      button.addEventListener('click', (event) => {
-        this._toggleDrawer(event, drawer);
-      });
-  
-      content.addEventListener('click', (event) => {
-        this._closeDrawer(event, drawer);
-      });
-    },
-  
-    _toggleDrawer(event, drawer) {
-      event.stopPropagation();
-      drawer.classList.toggle('open');
-    },
-  
-    _closeDrawer(event, drawer) {
-      event.stopPropagation();
-      drawer.classList.remove('open');
-    },
-  };
-  
-  export default DrawerInitiator;
