--- starter-project-harma/src/views/app.js
@@ -1,34 +1,0 @@
-import DrawerInitiator from '../utils/drawer-initiator';
-
-class App {
-    constructor({ button, drawer, content }) {
-      this._button = button;
-      this._drawer = drawer;
-      this._content = content;
-  
-      this._initialAppShell();
-    }
-  
-    _initialAppShell() {
-      DrawerInitiator.init({
-        button: this._button,
-        drawer: this._drawer,
-        content: this._content,
-      });
-    }
-  
-    async renderPage() {
-      const url = UrlParser.parseActiveUrlWithCombiner();
-      const page = routes[url];
-      this._content.innerHTML = await page.render();
-      await page.afterRender();
-      const mainContent = document.querySelector('#mainContent');
-      skipLink.addEventListener('click', (e) => {
-        e.preventDefault();
-        mainContent.scrollIntoView({ behavior: 'smooth' });
-        skipLink.blur();
-      });
-    }
-  }
-
-  export default App;
